[FEAT] #182 MainContentLayer 분리
